requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.1.0
Authlib==1.3.1
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==2.0.12
click==8.1.7
cloudpathlib==0.18.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.2
executing==2.1.0
faiss-gpu==1.7.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
grpcio==1.65.5
grpcio-health-checking==1.65.5
grpcio-tools==1.65.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
ipykernel==6.29.5
ipython==8.28.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.98
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
orjson==3.10.6
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pipx==1.6.0
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
rdflib==7.0.0
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.4
triton==3.0.0
typer==0.12.3
tabulate==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
userpath==1.9.2
validators==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weaviate==0.1.2
weaviate-client==4.9.2
Werkzeug==3.0.4
Whoosh==2.7.4
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
